# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API Documentation for Social App",
        "contact": {
            "email": "thawzinsoe.dev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "08b89b97682f31e515fb23aba1df2be0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "username",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "profile_pic": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "8a8ad725f7d674b065fdb10e3209e386",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "74e3fa2ca4e0898425f94c36ed1fa1b1",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "403": {
                        "description": "Not logged in"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/news-feed": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get news feed posts",
                "operationId": "8eb86e3558034196c713171e9c7c212e",
                "responses": {
                    "200": {
                        "description": "List of posts"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get my profile info",
                "operationId": "649eb25ad5fd26ff07839e616568cd45",
                "responses": {
                    "200": {
                        "description": "Profile details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get my posts",
                "operationId": "9ef08d91885a6eaab5be2c7401578056",
                "responses": {
                    "200": {
                        "description": "List of my posts"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/post": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create new post",
                "operationId": "53404509008b4b20cfe6b54e9e57b3f9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "image_file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "video_file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/post/{id}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete my post",
                "operationId": "b6a1dbf79d71bbfe0930223a985aefb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/react": {
            "post": {
                "tags": [
                    "Reacts"
                ],
                "summary": "Toggle like/react",
                "operationId": "92f2a9b2b80a75cde96b1fa6bf857deb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "React toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add comment",
                "operationId": "3e17dda440bfd541cb29cada576033e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "comment"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete my comment",
                "operationId": "0c3f049e4705ad470990e3ecde7a9787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Posts",
            "description": "Posts"
        },
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Reacts",
            "description": "Reacts"
        },
        {
            "name": "Comments",
            "description": "Comments"
        }
    ]
}